+++ commands/_general/vars.js
@@ -6,6 +6,7 @@ module.exports = {
    reecoom: 'https://www.nautiljon.com/images/perso/00/15/reecom_5851.jpg',
    githubUrl: 'https://github.com/fivekage/muslimbot',
    topggUrl: 'https://top.gg/bot/1183399354166415481',
+   paypalDonationUrl: 'https://www.paypal.com/donate/?hosted_button_id=T94D4G4PY3YT8',
    prayerGif: 'https://i.giphy.com/12ihpr4WmwKJsQ.gif',
    footerText: `${require('../../package.json').version} - MuslimBot 🕋 - For any help type /help command`
 };