+++ commands/religion/subscribe.js
@@ -4,30 +4,40 @@ const {
 const { usersModel, subscriptionsModel } = require('../../data/models.js');
 const logger = require('../../utils/logger.js');
 const { retrievePrayersOfTheDay } = require('../../utils/retrieve_prayers.js');
+const { CountriesAPI } = require('../../apis/countries_api.js');
 const vars = require('../_general/vars.js');
+
+const countriesAPI = new CountriesAPI();
+(async () => {
+   await countriesAPI.initialize(); // Pre-fetch countries data
+})();
+const COUNTRY_PARAM_NAME = 'country';
+const CITY_PARAM_NAME = 'city';
 
 module.exports.help = {
    name: 'subscribe',
    description: 'Subscription to get notifications for each prayer of the day according to the desired city',
    options: [
       {
-         name: 'city',
-         description: 'The city you are in',
-         type: 3,
-         required: true,
-      },
-      {
-         name: 'country',
+         name: COUNTRY_PARAM_NAME,
          description: 'The country you want to know the prayer times',
          type: 3,
          required: true,
+         autocomplete: true
+      },
+      {
+         name: CITY_PARAM_NAME,
+         description: 'The city you are in (if you don\'t find your city, you can write it, it will work)',
+         type: 3,
+         required: true,
+         autocomplete: true
       },
    ],
 };
 
 module.exports.run = async (_client, interaction) => {
-   const queryCountry = interaction.options.getString('country');
-   const queryCity = interaction.options.getString('city');
+   const queryCountry = interaction.options.getString(COUNTRY_PARAM_NAME);
+   const queryCity = interaction.options.getString(CITY_PARAM_NAME);
    const city = queryCity.charAt(0).toUpperCase() + queryCity.slice(1).toLowerCase();
    const country = queryCountry.charAt(0).toUpperCase() + queryCountry.slice(1).toLowerCase();
 
@@ -124,6 +134,40 @@ module.exports.run = async (_client, interaction) => {
       logger.error(e);
       await interaction.editReply({ content: 'Confirmation not received within 1 minute, cancelling', components: [] });
    }
+};
+
+module.exports.autocomplete = async (interaction) => {
+   const focusedOption = interaction.options.getFocused(true);
+   const focusedValue = interaction.options.getFocused();
+   let choices;
+
+   if (focusedOption.name === COUNTRY_PARAM_NAME) {
+      choices = countriesAPI.getCountriesName();
+   }
+   else if (focusedOption.name === CITY_PARAM_NAME) {
+      // Get country value
+      const countryName = interaction.options.getString(COUNTRY_PARAM_NAME);
+      if (countryName === null) {
+         choices = [];
+      } else {
+         choices = countriesAPI.getCountryCities(countryName);
+      }
+   }
+
+   // Ensure focused value is processed correctly (case-insensitive search)
+   const filtered = choices
+      .filter(choice => {
+         if (focusedValue === null) return true;
+         return choice.toLowerCase().startsWith(focusedValue.toLowerCase())
+      })
+      .slice(0, 25); // Limit to 25 suggestions
+
+   // Respond to the interaction within 3 seconds
+   await interaction.respond(
+      filtered.map(choice => ({ name: choice, value: choice }))
+   ).catch(error => {
+      throw new Error('Error responding to autocomplete interaction:', error);
+   });
 };
 
 module.exports.checkIfLocationExists = async (city, country) => {